+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the create-package project will be documented in this fil
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!--## [Unreleased]-->
+## [2.4.0] - 2019-04-04
 ### Added
 - Latest package version numbers in package.json will be automatically fetched.
 - Debug package and debug statements.
@@ -98,7 +99,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Script that deploys the templates into a project folder.
 - Setup for this project.
 
-[Unreleased]: https://github.com/Ionaru/create-package/compare/2.3.0...HEAD
+[Unreleased]: https://github.com/Ionaru/create-package/compare/2.4.0...HEAD
+[2.4.0]: https://github.com/Ionaru/create-package/compare/2.3.0...2.4.0
 [2.3.0]: https://github.com/Ionaru/create-package/compare/2.2.0...2.3.0
 [2.2.0]: https://github.com/Ionaru/create-package/compare/2.1.1...2.2.0
 [2.1.1]: https://github.com/Ionaru/create-package/compare/2.1.0...2.1.1